Fixed upload and moved setPosition()
